+++ src/main/main/PizzaFromage.java
@@ -2,6 +2,9 @@
  * @author bruel (from O'Reilly Head-First series)
  * @depend - * - Pizza
  */
+
+package main;
+
 public class PizzaFromage extends Pizza {
 
 }
