+++ src/main/main/PizzaFromageStyleBrest.java
@@ -1,14 +1,15 @@
+package main;
 
-public class PizzaFromageStyleStrasbourg extends Pizza{
+public class PizzaFromageStyleBrest extends Pizza{
 	
 	@Override
 	public String getNom() {
-		return "Pizza pate style Strasbourg et fromage";
+		return "Pizza sauce style brest et fromage";
 	}
 
 	@Override
 	public void preparer() {
-		this.garnitures.add("Mozzarella en lamelles");
+		this.garnitures.add("Parmigiano reggiano rape");
 		super.preparer();
 	}
 
@@ -19,7 +20,7 @@ public class PizzaFromageStyleStrasbourg extends Pizza{
 
 	@Override
 	public void couper() {
-		super.couper();
+		System.out.println("Decoupage en parts triangulaires");
 	}
 
 	@Override
