+++ src/main/main/PizzaGrecque.java
@@ -2,6 +2,9 @@
  * @author bruel (from O'Reilly Head-First series)
  * @depend - * - Pizza
  */
-public class PizzaPoivrons extends Pizza {
+
+package main;
+
+public class PizzaGrecque extends Pizza {
 
 }
