+++ src/main/main/PizzaPoivrons.java
@@ -2,6 +2,8 @@
  * @author bruel (from O'Reilly Head-First series)
  * @depend - * - Pizza
  */
-public class PizzaGrecque extends Pizza {
+package main;
+
+public class PizzaPoivrons extends Pizza {
 
 }
