+++ src/main/main/PizzeriaBrest.java
@@ -1,13 +1,14 @@
+package main;
 
-public class PizzeriaStrasbourg extends Pizzeria {
-	
+public class PizzeriaBrest extends Pizzeria {
+
 	@Override
 	public Pizza commanderPizza(String type) {
 
 		Pizza pizza;
 
 		if (type.equals("fromage")) {
-			pizza = new PizzaFromageStyleStrasbourg();
+			pizza = new PizzaFromageStyleBrest();
 		} else if (type.equals("grecque")) {
 			pizza = new PizzaGrecque();
 		} else {
